Code clean up.

+++ libcfile/libcfile_file.c
@@ -1802,7 +1802,7 @@ ssize_t libcfile_file_read_buffer_with_error_code(
 	size_t buffer_offset                    = 0;
 	size_t read_size                        = 0;
 	size_t read_size_remainder              = 0;
-	DWORD read_count                        = 0;
+	ssize_t read_count                      = 0;
 	BOOL result                             = FALSE;
 
 	if( file == NULL )
@@ -2080,7 +2080,7 @@ ssize_t libcfile_file_read_buffer_with_error_code(
 	{
 		/* The read was cut short
 		 */
-		if( read_count != (size_t) read_size )
+		if( read_count != (ssize_t) read_size )
 		{
 			return( (ssize_t) buffer_offset );
 		}
